+++ angular/projects/lib/src/lib/gridstack-item.component.ts
@@ -1,5 +1,5 @@
 /**
- * gridstack-item.component.ts 8.4.0-dev
+ * gridstack-item.component.ts 9.0.0
  * Copyright (c) 2022 Alain Dumesny - see GridStack root license
  */