+++ src/dd-base-impl.ts
@@ -1,5 +1,5 @@
 /**
- * dd-base-impl.ts 8.4.0-dev
+ * dd-base-impl.ts 9.0.0
  * Copyright (c) 2021-2022 Alain Dumesny - see GridStack root license
  */