+++ src/dd-element.ts
@@ -1,5 +1,5 @@
 /**
- * dd-elements.ts 8.4.0-dev
+ * dd-elements.ts 9.0.0
  * Copyright (c) 2021 Alain Dumesny - see GridStack root license
  */