+++ src/gridstack.ts
@@ -1,5 +1,5 @@
 /*!
- * GridStack 8.4.0-dev
+ * GridStack 9.0.0
  * https://gridstackjs.com/
  *
  * Copyright (c) 2021-2022 Alain Dumesny
@@ -1739,7 +1739,7 @@ export class GridStack {
     return this;
   }
 
-  static GDRev = '8.4.0-dev';
+  static GDRev = '9.0.0';
 
   /* ===========================================================================================
    * drag&drop methods that used to be stubbed out and implemented in dd-gridstack.ts